+++ packages/accounts/src/index.ts
@@ -32,12 +32,14 @@ export { MultiOwnerModularAccountFactoryAbi } from "./msca/abis/MultiOwnerModula
 export { UpgradeableModularAccountAbi } from "./msca/abis/UpgradeableModularAccount.js";
 export type * from "./msca/account-loupe/decorator.js";
 export { accountLoupeActions } from "./msca/account-loupe/decorator.js";
+export type * from "./msca/account-loupe/types.js";
 export type * from "./msca/account/multiOwnerAccount.js";
 export { createMultiOwnerModularAccount } from "./msca/account/multiOwnerAccount.js";
 export { standardExecutor } from "./msca/account/standardExecutor.js";
 export { createMultiOwnerModularAccountClient } from "./msca/client.js";
 export type * from "./msca/plugin-manager/decorator.js";
 export { pluginManagerActions } from "./msca/plugin-manager/decorator.js";
+export { installPlugin } from "./msca/plugin-manager/installPlugin.js";
 export type * from "./msca/plugins/multi-owner/index.js";
 export { multiOwnerPluginActions } from "./msca/plugins/multi-owner/index.js";
 export {