+++ packages/alchemy/src/signer/client/index.ts
@@ -37,10 +37,11 @@ export type AlchemySignerClientParams = z.input<
 // use this to override the parent org id for suborgs
 // used for whoami passkey requests to figure out the
 // suborg for a user
-const ROOT_ORG_ID =
-  process.env.ROOT_ORG_ID ??
-  process.env.NEXT_PUBLIC_ROOT_ORG_ID ??
-  "24c1acf5-810f-41e0-a503-d5d13fa8e830";
+const _ENV_ROOT_ORG_ID =
+  typeof process !== "undefined"
+    ? process.env.ROOT_ORG_ID ?? process.env.NEXT_PUBLIC_ROOT_ORG_ID
+    : null;
+const ROOT_ORG_ID = _ENV_ROOT_ORG_ID ?? "24c1acf5-810f-41e0-a503-d5d13fa8e830";
 
 /**
  * A lower level client used by the AlchemySigner used to communicate with
@@ -89,22 +90,32 @@ export class AlchemySignerClient {
         email,
         targetPublicKey: publicKey,
         expirationSeconds,
+        redirectParams: params.redirectParams?.toString(),
       });
 
       return response;
     }
 
+    // Passkey account creation flow
     const { attestation, challenge } = await this.getWebAuthnAttestation(
       params.creationOpts,
       { username: params.username }
     );
 
-    return this.request("/v1/signup", {
+    const result = await this.request("/v1/signup", {
       passkey: {
         challenge: base64UrlEncode(challenge),
         attestation,
       },
     });
+
+    this.user = {
+      orgId: result.orgId,
+      address: result.address!,
+      userId: result.userId!,
+    };
+
+    return result;
   };
 
   public initEmailAuth = async (
@@ -117,6 +128,7 @@ export class AlchemySignerClient {
       email,
       targetPublicKey: publicKey,
       expirationSeconds,
+      redirectParams: params.redirectParams?.toString(),
     });
   };