+++ site/snippets/getting-started/send-user-operation.ts
@@ -2,12 +2,12 @@ import { createModularAccountAlchemyClient } from "@alchemy/aa-alchemy";
 import {
   Address,
   LocalAccountSigner,
-  polygonMumbai,
+  sepolia,
   type Hex,
 } from "@alchemy/aa-core";
 
 // select your chain from @alchemy/aa-core
-const chain = polygonMumbai;
+const chain = sepolia;
 
 // The private key of your EOA that will be the signer to connect with the Modular Account
 const PRIVATE_KEY = "0xYourEOAPrivateKey" as Hex;
@@ -15,8 +15,8 @@ const signer = LocalAccountSigner.privateKeyToAccountSigner(PRIVATE_KEY);
 
 // [!code focus:22]
 (async () => {
-  // Create a provider to send user operations from your smart account
-  const provider = await createModularAccountAlchemyClient({
+  // Create a smart account client to send user operations from your smart account
+  const client = await createModularAccountAlchemyClient({
     // get your Alchemy API key at https://dashboard.alchemy.com
     apiKey: "ALCHEMY_API_KEY",
     chain,
@@ -25,12 +25,12 @@ const signer = LocalAccountSigner.privateKeyToAccountSigner(PRIVATE_KEY);
 
   // Fund your account address with ETH to send for the user operations
   // (e.g. Get Sepolia ETH at https://sepoliafaucet.com)
-  console.log("Smart Account Address: ", provider.getAddress()); // Log the smart account address
+  console.log("Smart Account Address: ", client.getAddress()); // Log the smart account address
 
   const vitalikAddress =
     "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045" as Address;
   // Send a user operation from your smart account to Vitalik that does nothing
-  const { hash: uoHash } = await provider.sendUserOperation({
+  const { hash: uoHash } = await client.sendUserOperation({
     uo: {
       target: vitalikAddress, // The desired target contract address
       data: "0x", // The desired call data
@@ -41,7 +41,7 @@ const signer = LocalAccountSigner.privateKeyToAccountSigner(PRIVATE_KEY);
   console.log("UserOperation Hash: ", uoHash); // Log the user operation hash
 
   // Wait for the user operation to be mined
-  const txHash = await provider.waitForUserOperationTransaction({
+  const txHash = await client.waitForUserOperationTransaction({
     hash: uoHash,
   });